app[REFACT] moved editPwdFetch in VM
