+++ packages/app/src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from "./useFetchAuth";
 export * from "./useTextInput";
 export * as signInScreenVM from "./signInScreenVM";
 export * as signUpScreenVM from "./signUpScreenVM";
+export * as passwordScreenVM from "./passwordScreenVM";
